Refactored feed and profile
