+++ src/components/Feed/components/TaskUnit/components/CompletedTask/index.js
@@ -1,7 +1,7 @@
 import * as React from "react";
-import "../Feed.scss";
-import "../Feed.css";
-import "../../../App.scss";
+import "../../../../Feed.scss";
+import "../../../../Feed.css";
+import "../../../../../../App.scss";
 import moment from "moment";
 
 class CompletedTask extends React.Component {
