+++ src/components/Feed/components/TaskUnit/components/SocialUnit/actions.js
@@ -1,0 +1,14 @@
+export const addClap = (info) => {
+  return {
+    type: "ADD_CLAP",
+    payload: info,
+  };
+};
+
+export const addMango = (info) => {
+  return {
+    type: "ADD_MANGO",
+    payload: info,
+  };
+};
+
