+++ src/components/Feed/components/TaskUnit/components/SocialUnit/components/MangoPopover/actions.js
@@ -1,0 +1,7 @@
+export const addMango = (info) => {
+  return {
+    type: "ADD_MANGO",
+    payload: info,
+  };
+};
+
