+++ src/components/UserProfile/components/Description/index.js
@@ -1,7 +1,7 @@
 import * as React from "react";
-import "../../../App.css";
-import "../../../App.scss";
-import "../UserProfile.css";
+import "../../../../App.css";
+import "../../../../App.scss";
+import "../../UserProfile.css";
 
 class UserDescription extends React.Component {
   render() {
